StudyVC Changes
